refactor: split large components into focused files
CraneDashboard.tsx (1321 lines) split into 10 focused modules:
- dashboard/MonitorFeed.tsx — CRT monitor wrapper with shake/VFX
- dashboard/TwistlockCam.tsx — merged L/R cams via `side` prop (eliminates duplication)
- dashboard/DownwardSpreaderCam.tsx — top-down alignment cam
- dashboard/WinchCam.tsx — cable drum + tension gauge
- dashboard/ThermalCam.tsx — thermal imaging with weather-reactive hotspots
- dashboard/Arctic360Cam.tsx — 360° arctic view (useMemo for stable random arrays)
- dashboard/TelemetryGraph.tsx — real-time bar chart
- dashboard/SystemStatusMonitor.tsx — system health grid
- dashboard/WeatherMonitor.tsx — arctic weather display
- dashboard/HeatedWindow.tsx — frost/defrost window simulation

CraneControls.tsx (478 lines) split:
- controls/useCranePhysics.ts — RAF physics loop + keyboard handlers
- controls/JoystickControl.tsx — reusable virtual joystick widget

UpgradeMenu.tsx (585 lines):
- upgrade/upgradeConfigs.ts — UPGRADE_CONFIGS, labels, colors extracted
- Fixed: inline document.createElement('style') moved to App.css
- Removed: leftover console.log

App.tsx (258 lines):
- components/MainMenu.tsx — extracted main menu component
- components/LoadingScreen.tsx — extracted loading screen component

Bug fixes: TwistlockCam duplication eliminated; Arctic360Cam Math.random()
calls stabilized with useMemo; HeatedWindow moved before its usage site.
